+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 0.28.6 (2025-05-01)
+
+
+### Experimental
+
+These changes are experimental and only available through the use of feature flags
+
+* HOST_CONFIGURATION_FEATURE - Prevent Windows ACL inheritance for host config script and log and grant full control to Administrators ([`98dfbfc`](https://github.com/aws-deadline/deadline-cloud-worker-agent/commit/98dfbfc9dbebdb02145c8413a76fe82f45635c58))
+* HOST_CONFIGURATION_FEATURE - delimit host configuration script with banners in worker logs and shutdown host on failure ([`f08bca8`](https://github.com/aws-deadline/deadline-cloud-worker-agent/commit/f08bca8656488d9af992b1dca4b4969f5bdf96b6))
+
+
+## 0.28.5 (2025-04-11)
+
+
+### Experimental
+
+These changes are experimental and only available through the use of feature flags
+
+* HOST_CONFIGURATION_FEATURE - run admin host configuration scripts once worker becomes STARTED (#601) ([`d925c65`](https://github.com/aws-deadline/deadline-cloud-worker-agent/commit/d925c65a052579675443d004295c2b70f019fe9b))
+
+
 ## 0.28.4 (2025-04-03)